gg/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
wikipedia==1.4.0
PyPDF2==3.0.1
numpy>=1.26.0
tqdm==4.66.1
pdfplumber==0.10.3
langchain-text-splitters==0.0.1
google-generativeai==0.3.2 